Adding bootstrap image to AlertsDLX. Abstracting out block to component.
